+++ scrapy_poet/utils.py
@@ -14,7 +14,7 @@ from web_poet.pages import ItemPage, is_injectable
 from scrapy_poet.providers import (
     providers,
 )
-from scrapy_poet.page_input_provider import PageObjectInputProvider
+from scrapy_poet.page_input_providers import PageObjectInputProvider
 
 _CALLBACK_FOR_MARKER = '__scrapy_poet_callback'
