+++ tests/test_middleware.py
@@ -13,7 +13,7 @@ import attr
 from scrapy_poet import callback_for
 from web_poet.pages import WebPage, ItemWebPage
 from scrapy_poet.providers import provides
-from scrapy_poet.page_input_provider import PageObjectInputProvider
+from scrapy_poet.page_input_providers import PageObjectInputProvider
 from web_poet.page_inputs import ResponseData
 from scrapy_poet.utils import DummyResponse
 from tests.utils import HtmlResource, crawl_items, capture_exceptions, \
