+++ tests/test_utils.py
@@ -8,7 +8,7 @@ from scrapy_poet.providers import (
     ResponseDataProvider,
     provides,
 )
-from scrapy_poet.page_input_provider import PageObjectInputProvider
+from scrapy_poet.page_input_providers import PageObjectInputProvider
 from web_poet.pages import ItemPage, WebPage
 
 from scrapy_poet.utils import (
